Relocate IDialogService + dialog support types into Gum.Presentation (ADR-0005 Phase 3)
Completes the relocation of the original-8 core tool interfaces into the headless Gum.Presentation assembly. Moves the framework-neutral dialog contract and its supporting value types out of the WPF shell, leaving the WPF-bound DialogService impl, IDialogServiceExt, concrete dialog VMs, and views in Gum.

Moved into Tools/Gum.Presentation/Dialogs/ (namespace Gum.Services.Dialogs preserved, zero consumer churn): IDialogService (split out of DialogService.cs), DialogViewModel (dead 'using Microsoft.Xaml.Behaviors.Core' stripped), MessageDialog (MessageDialogResult + MessageDialogStyle), OpenFileDialogOptions, SaveFileDialogOptions, and GetUserStringOptions (extracted from GetUserStringDialogViewModel.cs; the VM stays in the shell).

Gum.Presentation gains its first PackageReference, CommunityToolkit.Mvvm 8.4.0 (framework-neutral, matches Gum.csproj), required by DialogViewModel's RelayCommand.

DialogWindow.xaml's DataTemplate for the base DialogViewModel now needs an assembly-qualified xmlns ('presentation' -> Gum.Presentation) since that type crossed the assembly boundary; the abstract base is never resolved by DialogViewResolver, so runtime dialog resolution is unchanged.

Verified: Gum.Presentation builds standalone (no WPF/WinForms leak), GumFull.sln builds, GumToolUnitTests 998 passed / 4 skipped / 0 failed.
