+++ Tools/Gum.Presentation/Dialogs/GetUserStringOptions.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Gum.Services.Dialogs;
+
+public class GetUserStringOptions
+{
+    public Func<string?,string?>? Validator { get; set; }
+    public string? InitialValue { get; set; }
+    public bool PreSelect { get; set; }
+}
