+++ Tools/Gum.Presentation/Gum.Presentation.csproj
@@ -22,6 +22,12 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<!-- DialogViewModel uses RelayCommand. CommunityToolkit.Mvvm is framework-neutral
+		     (no WPF/WinForms), so it does not violate the headless boundary above. -->
+		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.0" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<ProjectReference Include="..\Gum.ProjectServices\Gum.ProjectServices.csproj" />
 	</ItemGroup>
